Refactor into library
