+++ Cargo.toml
@@ -6,8 +6,14 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-surf = "2.3"
 futures-lite = "*"
 serde = "1"
 serde_json = "*"
 dirs = "*"
+itertools = "*"
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+surf = { version = "2.3", default-features = false, features = ["wasm-client", "encoding"] }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+surf = { version = "2.3" }
