+++ src/lib.rs
@@ -1,0 +1,2 @@
+pub mod fuel_client;
+pub use fuel_client::*;
